lambda/requirements.txt
# Minimal dependencies - no compilation needed
boto3>=1.34.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
# Removed pydantic - not essential for Lambda
# email-validator - not essential

